refactor components for Shared directory structure
