+++ frontend/tests/unit/components/Shared/ActionButton.test.js
@@ -1,6 +1,6 @@
 import { render, screen } from "@testing-library/vue";
 
-import ActionButton from "@/components/ActionButton.vue";
+import ActionButton from "@/components/Shared/ActionButton.vue";
 
 describe("ActionButton", () => {
   it("renders text", () => {
